Initial view-viewmodel separation
